+++ src/main/java/com/domisa/domisa_backend/dating/dto/DatingRefreshTimeResponse.java
@@ -1,0 +1,9 @@
+package com.domisa.domisa_backend.dating.dto;
+
+import java.time.LocalDateTime;
+
+public record DatingRefreshTimeResponse(
+	LocalDateTime refreshAvailableAt,
+	boolean canRefresh
+) {
+}
