+++ src/main/java/com/domisa/domisa_backend/user/service/UserService.java
@@ -4,18 +4,13 @@ import com.domisa.domisa_backend.global.exception.GlobalErrorCode;
 import com.domisa.domisa_backend.global.exception.GlobalException;
 import com.domisa.domisa_backend.global.s3.service.S3ObjectUrlService;
 import com.domisa.domisa_backend.profile.dto.MyIntroductionResponse;
-import com.domisa.domisa_backend.user.dto.DatingProfileResponse;
-import com.domisa.domisa_backend.profileimage.entity.ProfileImage;
 import com.domisa.domisa_backend.user.dto.ContactDTO;
-import com.domisa.domisa_backend.user.dto.DatingRefreshTimeResponse;
 import com.domisa.domisa_backend.user.dto.UserCookiesResponse;
 import com.domisa.domisa_backend.user.dto.UserMeResponse;
 import com.domisa.domisa_backend.user.dto.UserLikesReceivedResponse;
 import com.domisa.domisa_backend.user.dto.UserLikesSentResponse;
 import com.domisa.domisa_backend.user.entity.User;
 import com.domisa.domisa_backend.user.repository.UserRepository;
-import java.time.Duration;
-import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
@@ -27,8 +22,6 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class UserService {
 
-	private static final Duration REFRESH_INTERVAL = Duration.ofHours(2);
-
 	private final UserRepository userRepository;
 	private final S3ObjectUrlService s3ObjectUrlService;
 
@@ -37,7 +30,6 @@ public class UserService {
 	public UserMeResponse getMe(User authUser) {
 		// 내 정보 조회는 프로필 이미지까지 함께 읽어서 응답한다.
 		User user = getRequiredUser(authUser);
-		ProfileImage profileImage = user.getProfileImage();
 
 		return new UserMeResponse(
 			user.getId(),
@@ -63,50 +55,6 @@ public class UserService {
 	public UserCookiesResponse getMyCookies(User authUser) {
 		User user = getRequiredUser(authUser);
 		return new UserCookiesResponse(user.getCookies() == null ? 0L : user.getCookies());
-	}
-
-	@Transactional(readOnly = true)
-	public DatingRefreshTimeResponse getDatingRefreshTime(User authUser) {
-		User user = getRequiredUser(authUser);
-		LocalDateTime refreshAvailableAt = user.getRefreshAt() == null
-			? null
-			: user.getRefreshAt().plus(REFRESH_INTERVAL);
-		boolean canRefresh = refreshAvailableAt == null || !LocalDateTime.now().isBefore(refreshAvailableAt);
-		return new DatingRefreshTimeResponse(refreshAvailableAt, canRefresh);
-	}
-
-	@Transactional(readOnly = true)
-	public DatingProfileResponse getDatingProfile(User authUser, Long userId) {
-		User requester = getRequiredUser(authUser);
-		User targetUser = userRepository.findDatingProfileById(userId)
-			.orElseThrow(() -> new GlobalException(GlobalErrorCode.USER_NOT_FOUND));
-
-		boolean isBlurred = requester.getMyBlurs() == null || !requester.getMyBlurs().contains(userId);
-		boolean hasSentLike = requester.getMyTypes() != null && requester.getMyTypes().contains(userId);
-
-		String profileUrl = isBlurred
-			? s3ObjectUrlService.getOriginBlurUrl(targetUser.getProfileImage())
-			: s3ObjectUrlService.getProfileImageUrl(targetUser.getProfileImage());
-		ContactDTO contact = isBlurred
-			? null
-			: new ContactDTO(targetUser.getContactType(), targetUser.getContact());
-
-		return new DatingProfileResponse(
-			targetUser.getId(),
-			targetUser.getNickname(),
-			targetUser.getAge(),
-			targetUser.getAnimalProfile(),
-			profileUrl,
-			targetUser.getIntroduction() == null ? null : targetUser.getIntroduction().getQ1(),
-			targetUser.getIntroduction() == null ? null : targetUser.getIntroduction().getQ2(),
-			targetUser.getIntroduction() == null ? null : targetUser.getIntroduction().getQ3(),
-			targetUser.getCard() == null ? null : targetUser.getCard().getDatingStyle(),
-			targetUser.getCard() == null ? null : targetUser.getCard().getIdealType(),
-			targetUser.getCard() == null ? null : targetUser.getCard().getMbti(),
-			contact,
-			isBlurred,
-			hasSentLike
-		);
 	}
 
 	@Transactional(readOnly = true)
